requirements.txt

torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
Pillow>=8.0.0
opencv-python>=4.5.0
tqdm>=4.60.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
streamlit>=1.0.0
requests>=2.25.0